feat: periodic pruning of payment tables

+++ api/clients/v2/disperser_client.go
@@ -139,7 +139,7 @@ func (c *disperserClient) DisperseBlob(
 	}
 
 	symbolLength := encoding.GetBlobLengthPowerOf2(uint(len(data)))
-	payment, err := c.accountant.AccountBlob(ctx, uint32(symbolLength), quorums, salt)
+	payment, err := c.accountant.AccountBlob(ctx, uint64(symbolLength), quorums, salt)
 	if err != nil {
 		return nil, [32]byte{}, fmt.Errorf("error accounting blob: %w", err)
 	}
